rswitch: allocated descriptors from reserved memory
This is breaking change that requires similar change in device tree.

It is needed because current rswitch GWCA implementation prevents proper
chain initialization with different OSIDs. As a workaround, we are need
to allocate LINKFIX table (aka base descriptors) exactly on the same
address for all domains. To achieve this, we need to provide and use
reserved memory page at a fixed address.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>