net: ethernet: renesas: rswitch: Optimize PF usage
In cases when user tries to setup TC filters that
overlaps neighboring bytes, we can reduce perfect
filter usage. For example: if user setups filtering
by source and destination IpV4 addresses, currently
we setup 2 four-byte filters in mask mode but with
this patch, it converts this case into 1 four-byte
filter in expand mode which performs in the same
way but saves one four-byte filter entry. Also,
in the case when user needs to filter packets by
one byte (e.g. proto type) we can use any kind of
filter with an appropriate mask to filter this and
with this patch, the algorithm will find the less
used filter to perform filtering that will balance
different perfect filters types usage.

Reviewed-by: Dmytro Firsov <dmytro_firsov@epam.com>
Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Signed-off-by: Dat Vo <dat.vo.yb@renesas.com>